first clean up just to see read me
