+++ app/README.md
@@ -1,1 +1,53 @@
-26.02 Created initial folder for the Flask Application for Project Krass. 
+# Understanding the code for the Krass App
+
+Folder Structure
+
+app/
+├── app.py                  ← Flask demo app — entry point
+├── main.py 
+├── pyproject.toml      <- used for initializing the dependencies 
+├── templates/
+│   ├── predict.html
+│   └── index.html          ← minimal HTML form for the demo
+├── functions/                
+│    ├── train.py                ← trains and saves all models — run once
+│    ├── scenarios.py            ← generates synthetic future data for predictions
+│    ├── splitter.py             ← data preparation and time series splitting
+│    ├── preprocessing.py        ← encoding, imputation and scaling
+│    ├── krr.py                  ← Kernel Ridge Regression model
+│    ├── assessment.py           ← MSE and R² scoring
+│    └── cross_validator.py      ← cross validation loop
+└── jupyter_test_arena/
+    ├── preprocessing_implementation.ipynb  ← all development code and tests
+    └── test_models/        ← models generated by the notebook — not used by app
+
+# How to run the app
+The app uses its own venv to prevent any error.
+
+`uv sync`
+`uv run python app.py`
+
+To retrain all models from scratch:
+
+`uv run python train.py`
+
+This will take a while. It runs a full grid search for each health variable.
+
+# Modulal descriptions 
+
+
+- splitter.py contains prepare_data() and the Splitter class. prepare_data() loads the dataframe, drops the non-target health variables and identifier columns, and returns X, y and years. Splitter takes those arrays and produces time series folds via time_series_splits() and the final train/test split via get_test_split(). This is the first thing called in the pipeline.
+- preprocessing.py contains detect_categorical_columns(), SimpleImputer, StandardScaler, OneHotEncoder and Preprocessor. These are independent classes connected through composition inside Preprocessor,  not through inheritance. Preprocessor is the only class called from outside this file. It applies encoding → imputation → scaling in that order. The fit/transform separation is strictly enforced — fit_transform() is only ever called on training data.
+- krr.py  contains gaussian_kernel() and KernelRidgeRegression. 
+- assessment.py contains Assessment with two methods  mean_squared_error() and r2_score(). This class was created by Sanjeev as well and it  computes scores and does not manage any CV loop or state. It is used only for final test set evaluation.
+- cross_validator.py contains CrossValidator which owns the full CV loop. It imports Preprocessor directly and creates a fresh instance per fold to prevent data leakage between folds. It calls Assessment internally for scoring and returns a dictionary with mse and r2 lists. start_fold controls which folds count toward hyperparameter selection.
+- train.py orchestrates the full pipeline for each health variable. It calls splitter.py, cross_validator.py, preprocessing.py, krr.py and assessment.py in sequence. After fitting the final model it saves both the KRR and the preprocessor together as a .pkl file via save_model(). It also contains load_model() which is imported by app.py.
+- scenarios.py TO BE UPDATED!!!contains the SCENARIOS dictionary and generate_scenario(). It takes the 2023 row for a selected county as a baseline, creates 10 synthetic (hypothetical for now) future rows and applies the corresponding climate changes. It returns a raw unscaled feature matrix that must go through preprocessor.transform() before prediction. County names are not unique across states so it uses state_abbr to uniquely identify a county (e.g. CLINTON COUNTY exists in Michigan, Iowa and New York). The app was initially crashing because I hadn't considered the possibility of duplicate names. Including the state fixes it, but we could also consider using the FIPS code if necessary to properly identify counties.
+- randon_fourier_features.py 
+- app.py is a minimal Flask app with two routes. GET / renders the form. POST /predict loads the model, generates scenario data, preprocesses and predicts. The full prediction flow is:
+
+
+load_model(target)              → krr, preprocessor  from train.py
+generate_scenario(...)          → X_future, years     from scenarios.py
+preprocessor.transform(X)      → X_scaled            from preprocessing.py
+krr.predict(X_scaled)          → y_pred              from krr.py
