+++ app/functions/scenarios.py
