+++ app/pyproject.toml
@@ -1,16 +1,11 @@
 [project]
-name = "app"
+name = "krass-app"
 version = "0.1.0"
-description = "Add your description here"
+description = "Creation of an app using Flask. We will be implemeting KRR and XGBoost."
 readme = "README.md"
-requires-python = ">=3.14"
+requires-python = ">=3.12"
 dependencies = [
     "flask>=3.1.3",
-    "ipykernel>=7.2.0",
-    "matplotlib>=3.10.8",
-    "numpy>=2.4.3",
-    "pandas>=3.0.1",
-    "scipy>=1.17.1",
-    "seaborn>=0.13.2",
-    "statsmodels>=0.14.6",
+    "numpy>=2.4.4",
+    "pandas>=3.0.2",
 ]
