+++ app_old/README.md
@@ -1,0 +1,1 @@
+26.02 Created initial folder for the Flask Application for Project Krass.
