+++ app_old/kernel_ridge.py
