+++ app_old/templates/index.html
@@ -1,0 +1,6 @@
+{% extends "base.html" %}
+{% block title %}Home — Climate & Health Explorer{% endblock %}
+{% block content %}
+<h1>Climate & Health Explorer</h1>
+<p>Welcome.</p>
+{% endblock %}
