+++ app_old/templates/predict.html
@@ -1,0 +1,6 @@
+{% extends "base.html" %}
+{% block title %}Predict — Climate & Health Explorer{% endblock %}
+{% block content %}
+<h1>Predict</h1>
+<p>Coming soon.</p>
+{% endblock %}
