--- flask_app_santiago/main.py
@@ -1,6 +1,0 @@
-def main():
-    print("Hello from flask-app-santiago!")
-
-
-if __name__ == "__main__":
-    main()
